+++ crates/types/src/bools.rs
@@ -1,5 +1,6 @@
 use crate::literals::LiteralValue;
-use crate::{SchemaType, Schematic};
+use crate::schema_type::SchemaType;
+use crate::Schematic;
 
 #[derive(Clone, Debug, Default)]
 pub struct BooleanType {
