+++ crates/types/src/schema.rs
@@ -1,4 +1,4 @@
-use crate::SchemaType;
+use crate::schema_type::SchemaType;
 
 #[derive(Clone, Debug, Default)]
 pub struct Schema {
